+++ CHANGELOG.md
@@ -17,6 +17,17 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+
+- `plugin-giphy`
+  - [#2312](https://github.com/wix/ricos/pull/2312) TS & Add missing hotkeys to giphy selector and more descriptive alt text
+
+- `fullscreen`
+  - [#3162](https://github.com/wix/ricos/pull/3162) videos pause when scrolled out of view
+  - [#3082](https://github.com/wix/ricos/pull/3082) image centered correctly on mobile
+
+
 </details>
 <hr/>